[AIRFLOW-YYYY] Move task functions from helpers to BaseOperator
