+++ airflow/models/baseoperator.py
@@ -26,7 +26,9 @@ import sys
 import warnings
 from abc import ABCMeta, abstractmethod
 from datetime import datetime, timedelta
-from typing import Any, Callable, ClassVar, Dict, FrozenSet, Iterable, List, Optional, Set, Tuple, Type, Union
+from typing import (
+    Any, Callable, ClassVar, Dict, FrozenSet, Iterable, List, Optional, Set, Tuple, Type, Union, cast,
+)
 
 import attr
 import jinja2
@@ -49,7 +51,7 @@ from airflow.ti_deps.deps.trigger_rule_dep import TriggerRuleDep
 from airflow.utils import timezone
 from airflow.utils.db import provide_session
 from airflow.utils.decorators import apply_defaults
-from airflow.utils.helpers import validate_key
+from airflow.utils.key_validation import validate_key
 from airflow.utils.log.logging_mixin import LoggingMixin
 from airflow.utils.operator_resources import Resources
 from airflow.utils.trigger_rule import TriggerRule
@@ -670,6 +672,7 @@ class BaseOperator(Operator, LoggingMixin):
 
         for k, v in self.__dict__.items():
             if k not in shallow_copy:
+                # noinspection PyArgumentList
                 setattr(result, k, copy.deepcopy(v, memo))
             else:
                 setattr(result, k, copy.copy(v))
@@ -862,6 +865,7 @@ class BaseOperator(Operator, LoggingMixin):
             tasks += [
                 t.task_id for t in self.get_flat_relatives(upstream=False)]
 
+        # noinspection PyUnresolvedReferences
         qry = qry.filter(TaskInstance.task_id.in_(tasks))
 
         count = qry.count()
@@ -1115,6 +1119,85 @@ class BaseOperator(Operator, LoggingMixin):
                 } | {'_task_type', 'subdag', 'ui_color', 'ui_fgcolor', 'template_fields'})
 
         return cls.__serialized_fields
+
+    @staticmethod
+    def chain(*tasks: Union['BaseOperator', List['BaseOperator']]):
+        r"""
+        Given a number of tasks, builds a dependency chain.
+        Support mix airflow.models.BaseOperator and List[airflow.models.BaseOperator].
+        If you want to chain between two List[airflow.models.BaseOperator], have to
+        make sure they have same length.
+
+        chain(t1, [t2, t3], [t4, t5], t6)
+
+        is equivalent to
+
+          / -> t2 -> t4 \
+        t1               -> t6
+          \ -> t3 -> t5 /
+
+        t1.set_downstream(t2)
+        t1.set_downstream(t3)
+        t2.set_downstream(t4)
+        t3.set_downstream(t5)
+        t4.set_downstream(t6)
+        t5.set_downstream(t6)
+
+        :param tasks: List of tasks or List[airflow.models.BaseOperator] to set dependencies
+        :type tasks: List[airflow.models.BaseOperator] or airflow.models.BaseOperator
+        """
+        for index, up_task in enumerate(tasks[:-1]):
+            down_task = tasks[index + 1]
+            if isinstance(up_task, BaseOperator):
+                up_task.set_downstream(down_task)
+                continue
+            if isinstance(down_task, BaseOperator):
+                down_task.set_upstream(up_task)
+                continue
+            if not isinstance(up_task, List) or not isinstance(down_task, List):
+                raise TypeError(
+                    'Chain not supported between instances of {up_type} and {down_type}'.format(
+                        up_type=type(up_task), down_type=type(down_task)))
+            up_task_list = cast(List[BaseOperator], up_task)
+            down_task_list = cast(List[BaseOperator], down_task)
+            if len(up_task_list) != len(down_task_list):
+                raise AirflowException(
+                    f'Chain not supported different length Iterable '
+                    f'but get {len(up_task_list)} and {len(down_task_list)}')
+            for up_t, down_t in zip(up_task_list, down_task_list):
+                up_t.set_downstream(down_t)
+
+    @staticmethod
+    def cross_downstream(from_tasks: List['BaseOperator'],
+                         to_tasks: Union['BaseOperator', List['BaseOperator']]):
+        r"""
+        Set downstream dependencies for all tasks in from_tasks to all tasks in to_tasks.
+        E.g.: cross_downstream(from_tasks=[t1, t2, t3], to_tasks=[t4, t5, t6])
+        Is equivalent to:
+
+        t1 --> t4
+           \ /
+        t2 -X> t5
+           / \
+        t3 --> t6
+
+        t1.set_downstream(t4)
+        t1.set_downstream(t5)
+        t1.set_downstream(t6)
+        t2.set_downstream(t4)
+        t2.set_downstream(t5)
+        t2.set_downstream(t6)
+        t3.set_downstream(t4)
+        t3.set_downstream(t5)
+        t3.set_downstream(t6)
+
+        :param from_tasks: List of tasks to start from.
+        :type from_tasks: List[airflow.models.BaseOperator]
+        :param to_tasks: List of tasks to set as downstream dependencies.
+        :type to_tasks: List[airflow.models.BaseOperator]
+        """
+        for task in from_tasks:
+            task.set_downstream(to_tasks)
 
 
 @attr.s(auto_attribs=True)
