+++ airflow/utils/key_validation.py
@@ -1,0 +1,41 @@
+# Licensed to the Apache Software Foundation (ASF) under one
+# or more contributor license agreements.  See the NOTICE file
+# distributed with this work for additional information
+# regarding copyright ownership.  The ASF licenses this file
+# to you under the Apache License, Version 2.0 (the
+# "License"); you may not use this file except in compliance
+# with the License.  You may obtain a copy of the License at
+#
+#   http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing,
+# software distributed under the License is distributed on an
+# "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+# KIND, either express or implied.  See the License for the
+# specific language governing permissions and limitations
+# under the License.
+
+"""Validates keys used in various places in Airflow code"""
+import re
+
+from airflow import AirflowException
+
+
+def validate_key(k, max_length=250):
+    """
+    Validates value used as a key.
+    """
+    if not isinstance(k, str):
+        raise TypeError("The key has to be a string")
+    elif len(k) > max_length:
+        raise AirflowException(
+            "The key has to be less than {0} characters".format(max_length))
+    elif not KEY_REGEX.match(k):
+        raise AirflowException(
+            "The key ({k}) has to be made of alphanumeric characters, dashes, "
+            "dots and underscores exclusively".format(k=k))
+    else:
+        return True
+
+
+KEY_REGEX = re.compile(r'^[\w.-]+$')
