+++ tests/models/test_baseoperator.py
@@ -17,15 +17,17 @@
 # specific language governing permissions and limitations
 # under the License.
 
-import datetime
 import unittest
 import uuid
+from datetime import date, datetime
 from unittest import mock
 
 import jinja2
 from parameterized import parameterized
 
+from airflow.exceptions import AirflowException
 from airflow.models import DAG
+from airflow.models.baseoperator import BaseOperator
 from airflow.operators.dummy_operator import DummyOperator
 from tests.models import DEFAULT_DATE
 from tests.test_utils.mock_operators import MockNamedTuple, MockOperator
@@ -81,8 +83,8 @@ class TestBaseOperator(unittest.TestCase):
                 {"foo": "bar"},
                 {"key_{{ foo }}_1": 1, "key_2": "bar_2"},
             ),
-            (datetime.date(2018, 12, 6), {"foo": "bar"}, datetime.date(2018, 12, 6)),
-            (datetime.datetime(2018, 12, 6, 10, 55), {"foo": "bar"}, datetime.datetime(2018, 12, 6, 10, 55)),
+            (date(2018, 12, 6), {"foo": "bar"}, date(2018, 12, 6)),
+            (datetime(2018, 12, 6, 10, 55), {"foo": "bar"}, datetime(2018, 12, 6, 10, 55)),
             (MockNamedTuple("{{ foo }}_1", "{{ foo }}_2"), {"foo": "bar"}, MockNamedTuple("bar_1", "bar_2")),
             ({"{{ foo }}_1", "{{ foo }}_2"}, {"foo": "bar"}, {"bar_1", "bar_2"}),
             (None, {}, None),
@@ -241,7 +243,44 @@ class TestBaseOperator(unittest.TestCase):
         task = DummyOperator(task_id="default-resources")
         self.assertIsNone(task.resources)
 
+    # noinspection PyUnresolvedReferences
     def test_custom_resources(self):
         task = DummyOperator(task_id="custom-resources", resources={"cpus": 1, "ram": 1024})
         self.assertEqual(task.resources.cpus.qty, 1)
         self.assertEqual(task.resources.ram.qty, 1024)
+
+    def test_cross_downstream(self):
+        """Test if all dependencies between tasks are all set correctly."""
+        dag = DAG(dag_id="test_dag", start_date=datetime.now())
+        start_tasks = [DummyOperator(task_id="t{i}".format(i=i), dag=dag) for i in range(1, 4)]
+        end_tasks = [DummyOperator(task_id="t{i}".format(i=i), dag=dag) for i in range(4, 7)]
+        BaseOperator.cross_downstream(from_tasks=start_tasks, to_tasks=end_tasks)
+
+        for start_task in start_tasks:
+            self.assertCountEqual(start_task.get_direct_relatives(upstream=False), end_tasks)
+
+    def test_chain(self):
+        dag = DAG(dag_id='test_chain', start_date=datetime.now())
+        [op1, op2, op3, op4, op5, op6] = [
+            DummyOperator(task_id='t{i}'.format(i=i), dag=dag)
+            for i in range(1, 7)
+        ]
+        BaseOperator.chain(op1, [op2, op3], [op4, op5], op6)
+
+        self.assertCountEqual([op2, op3], op1.get_direct_relatives(upstream=False))
+        self.assertEqual([op4], op2.get_direct_relatives(upstream=False))
+        self.assertEqual([op5], op3.get_direct_relatives(upstream=False))
+        self.assertCountEqual([op4, op5], op6.get_direct_relatives(upstream=True))
+
+    def test_chain_not_support_type(self):
+        dag = DAG(dag_id='test_chain', start_date=datetime.now())
+        [op1, op2] = [DummyOperator(task_id='t{i}'.format(i=i), dag=dag) for i in range(1, 3)]
+        with self.assertRaises(TypeError):
+            # noinspection PyTypeChecker
+            BaseOperator.chain([op1, op2], 1)
+
+    def test_chain_different_length_iterable(self):
+        dag = DAG(dag_id='test_chain', start_date=datetime.now())
+        [op1, op2, op3, op4, op5] = [DummyOperator(task_id='t{i}'.format(i=i), dag=dag) for i in range(1, 6)]
+        with self.assertRaises(AirflowException):
+            BaseOperator.chain([op1, op2], [op3, op4, op5])
